fix: BrowserAuthError Interaction_in_progress
